fix circul;ar dependencies
